PLAT-760: split HttpClient into an interface class and the HttpClientV1 implementation class
